requirements.txt
requests
selenium 
pyyaml
python-telegram-bot
2captcha-python